Update tx deltas

+++ sdk/transaction_model.go
@@ -735,12 +735,12 @@ func (dto *transferTransactionDTO) toStruct() (*TransferTransaction, error) {
 // ModifyMultisigAccountTransaction
 type ModifyMultisigAccountTransaction struct {
 	AbstractTransaction
-	MinApprovalDelta int32
-	MinRemovalDelta  int32
+	MinApprovalDelta int
+	MinRemovalDelta  int
 	Modifications    []*MultisigCosignatoryModification
 }
 
-func NewModifyMultisigAccountTransaction(deadline *Deadline, minApprovalDelta int32, minRemovalDelta int32, modifications []*MultisigCosignatoryModification, networkType NetworkType) (*ModifyMultisigAccountTransaction, error) {
+func NewModifyMultisigAccountTransaction(deadline *Deadline, minApprovalDelta int, minRemovalDelta int, modifications []*MultisigCosignatoryModification, networkType NetworkType) (*ModifyMultisigAccountTransaction, error) {
 	if modifications == nil {
 		return nil, errors.New("modifications must not be nil")
 	}
@@ -802,8 +802,8 @@ func (tx *ModifyMultisigAccountTransaction) generateBytes() ([]byte, error) {
 	transactions.ModifyMultisigAccountTransactionBufferStart(builder)
 	transactions.TransactionBufferAddSize(builder, 123+(33*len(tx.Modifications)))
 	tx.AbstractTransaction.buildVectors(builder, v, signatureV, signerV, deadlineV, fV)
-	transactions.ModifyMultisigAccountTransactionBufferAddMinApprovalDelta(builder, tx.MinApprovalDelta)
-	transactions.ModifyMultisigAccountTransactionBufferAddMinRemovalDelta(builder, tx.MinRemovalDelta)
+	transactions.ModifyMultisigAccountTransactionBufferAddMinApprovalDelta(builder, int32(tx.MinApprovalDelta))
+	transactions.ModifyMultisigAccountTransactionBufferAddMinRemovalDelta(builder, int32(tx.MinRemovalDelta))
 	transactions.ModifyMultisigAccountTransactionBufferAddNumModifications(builder, len(tx.Modifications))
 	transactions.ModifyMultisigAccountTransactionBufferAddModifications(builder, mV)
 	t := transactions.TransactionBufferEnd(builder)
@@ -815,8 +815,8 @@ func (tx *ModifyMultisigAccountTransaction) generateBytes() ([]byte, error) {
 type modifyMultisigAccountTransactionDTO struct {
 	Tx struct {
 		abstractTransactionDTO
-		MinApprovalDelta int32                                 `json:"minApprovalDelta"`
-		MinRemovalDelta  int32                                 `json:"minRemovalDelta"`
+		MinApprovalDelta int                                   `json:"minApprovalDelta"`
+		MinRemovalDelta  int                                   `json:"minRemovalDelta"`
 		Modifications    []*multisigCosignatoryModificationDTO `json:"modifications"`
 	} `json:"transaction"`
 	TDto transactionInfoDTO `json:"meta"`
